Demo-automation/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fabric-demo-automation"
version = "0.1.0"
description = "Automated Fabric Ontology Demo Setup Tool"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Azure Ontology Team"}
]
keywords = ["fabric", "ontology", "demo", "automation", "microsoft"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "azure-identity>=1.15.0",
    "azure-storage-file-datalake>=12.14.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "rdflib>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
fabric-demo = "demo_automation.cli:main"

[project.urls]
"Homepage" = "https://github.com/falloutxAY/rdf-dtdl-fabric-ontology-converter"
"Bug Tracker" = "https://github.com/falloutxAY/rdf-dtdl-fabric-ontology-converter/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
